==> 1. Microbial Interactions (MI) in a SynCom/2.SynCom_Guild_classification_Competition_prediction.Rmd <==
---
title: "SynCom Community Modifs Multivariate"
author: "Oriane Moyne"
date: "10/19/2021"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

## Load libraries

```{r, message = FALSE}
# Packages I commonly use
library(stringr)
library(ggplot2)
library(FactoMineR)
library(factoextra)
library(reshape)
library(reshape2)
library(gplots)
```

#--------------------------------

# Samples description

Multi-omics: metagenomics (metaG), metatranscriptomics (metaT), metaRibo-Seq (metaRS) analysis of 4-days (T1) and 7-days (T2) growth SynCom.  

- X1 and X2 samples: Full SynCom (16 strains)  
- X3: SynCom - _Lysobacter_ dropout
- X4: SynCom - _Mucilaginibacter_ dropout
- X5: SynCom - _Burkholderia_ dropout
- X6: SynCom - _Burkholderia_, _Rhizobium_, _Rhodocccus_, _Brevibacillus_ dropouts.

#--------------------------------

# Overall view of the data: Pathway Analysis

## Import KEGG-annotated data

Output from previous script.

```{r}
# Count tables with taxonomy annotations (Woltka gotu output)
syncom_KEGG <- read.table("/home/moyne/Moyne_disk/Soil_project/Final_datasets/Syncom_formatted_filtered_corrected_RPKM_KEGG_no_NA.tsv", 
                    sep="\t", 
                    dec=".", 
                    header = TRUE, 
                    stringsAsFactors = FALSE, 
                    na.strings = NA, 
                    check.names = FALSE, 
                    comment.char = "")

syncom_KEGG$genus_strain <- str_split(syncom_KEGG$genus_strain, "_", simplify = TRUE)[,1]

 # Filter out very low counts (optional)
#syncom_KEGG[, 9:ncol(syncom_KEGG)][syncom_KEGG[, 9:ncol(syncom_KEGG)]<1] <- 0
```

## Aggregate by pathway

```{r, eval = TRUE}
syncom_agg <- aggregate(syncom_KEGG[, 9:ncol(syncom_KEGG)], by = list(kegg = syncom_KEGG$pathway, genus_strain = syncom_KEGG$genus_strain), sum)
```


## Overall view of the samples
 
### PCA, all -omics

```{r}
# Data formatting
syncom_pca <- aggregate(syncom_KEGG[, 9:ncol(syncom_KEGG)], by = list(kegg = syncom_KEGG$pathway, genus_strain = syncom_KEGG$genus_strain), sum)
rownames(syncom_pca) <- paste(syncom_pca$genus_strain, syncom_pca$kegg, sep = "|")

syncom_pca <- syncom_pca[, 3:ncol(syncom_pca)]

syncom_pca <- data.frame(t(syncom_pca))

Community <- str_split(rownames(syncom_pca), "_", simplify = TRUE)[,1]
Replicate <- str_split(rownames(syncom_pca), "_", simplify = TRUE)[,2]
Time <- str_split(rownames(syncom_pca), "_", simplify = TRUE)[,3]
Omic <- str_split(rownames(syncom_pca), "_", simplify = TRUE)[,4]

syncom_pca <- data.frame(cbind(Community, Replicate, Time, Omic, syncom_pca))

syncom_pca <- syncom_pca[!str_detect(rownames(syncom_pca), "1_2_T2_metaG"), ]
```


```{r, eval = TRUE}
syncom_pca <- syncom_pca[syncom_pca$Time == "T2", ]

# metaG
pca_metaG <- PCA(log(syncom_pca[syncom_pca$Omic == "metaG", 5:ncol(syncom_pca)] + 1), 
          # quali.sup = c(1:4), 
           scale.unit = TRUE, 
           graph = FALSE
           )

fviz_pca_ind(pca_metaG,
             habillage = syncom_pca[syncom_pca$Omic == "metaG", ]$Community, 
             geom = c("point", "text"),              
             repel = TRUE
             )

# metaT
pca_metaT <- PCA(log(syncom_pca[syncom_pca$Omic == "metaT", 5:ncol(syncom_pca)] + 1), 
          # quali.sup = c(1:4), 
           scale.unit = TRUE, 
           graph = FALSE
           )

fviz_pca_ind(pca_metaT,
             habillage = syncom_pca[syncom_pca$Omic == "metaT", ]$Community, 
             geom = c("point", "text"),              
             repel = TRUE
             )

# metaRS
pca_metaRS <- PCA(log(syncom_pca[syncom_pca$Omic == "metaRS", 5:ncol(syncom_pca)] + 1), 
          # quali.sup = c(1:4), 
           scale.unit = TRUE, 
           graph = FALSE
           )

fviz_pca_ind(pca_metaRS,
             habillage = syncom_pca[syncom_pca$Omic == "metaRS", ]$Community, 
             geom = c("point", "text"),              
             repel = TRUE
             )
```

Data for all the -omics are very reproducible within conditions. Reminder:  


# TE calculation

```{r, eval = TRUE}
syncom_KEGG_id <- aggregate(syncom_KEGG[, 9:ncol(syncom_KEGG)], by = list(kegg = syncom_KEGG$pathway, genus_strain = syncom_KEGG$genus_strain), sum)

# Create separate matrices for each omic
metaG_mat <- syncom_KEGG_id[, str_detect(colnames(syncom_KEGG_id), "metaG")]
colnames(metaG_mat) <- str_replace(colnames(metaG_mat), "_metaG", "")
metaT_mat <- syncom_KEGG_id[, str_detect(colnames(syncom_KEGG_id), "metaT")]
colnames(metaT_mat) <- str_replace(colnames(metaT_mat), "_metaT", "")
metaRS_mat <- syncom_KEGG_id[, str_detect(colnames(syncom_KEGG_id), "metaRS")]
colnames(metaRS_mat) <- str_replace(colnames(metaRS_mat), "_metaRS", "")

# order them the same
metaRS_mat <- metaRS_mat[, colnames(metaT_mat)]
metaG_mat <- metaG_mat[, colnames(metaT_mat)]

# Rename columns
colnames(metaG_mat) <- paste(colnames(metaT_mat), "metaG", sep = "_")
colnames(metaT_mat) <- paste(colnames(metaT_mat), "metaT", sep = "_")
colnames(metaRS_mat) <- paste(colnames(metaRS_mat), "metaRS", sep = "_")

# Filter out low counts (< 10 RPKM for the whole pathway)
metaG_mat[metaG_mat<10] <- 0
metaRS_mat[metaRS_mat<10] <- 0
metaT_mat[metaT_mat<10] <- 0

TE_mat <- metaRS_mat / metaT_mat
colnames(TE_mat) <- str_replace(colnames(TE_mat), "_metaRS", "_TE")

# NaN produced when 0/0
TE_mat[is.na(TE_mat)] <- 0

# Inf produced when x/0
is.na(TE_mat) <- sapply(TE_mat, is.infinite)
TE_mat[is.na(TE_mat)] <- 0
#TE_mat[TE_mat<1] <- 0

TE_syncom <- data.frame(cbind(syncom_KEGG_id[, 1:2], TE_mat))
```

### Data formatting

```{r}
TE_pca <- TE_syncom
rownames(TE_pca) <- paste(TE_pca[,2], TE_pca[,1], sep = "|")

TE_pca <- TE_pca[, 3:ncol(TE_pca)]

TE_pca <- data.frame(t(TE_pca))

Community <- str_split(rownames(TE_pca), "_", simplify = TRUE)[,1]
Replicate <- str_split(rownames(TE_pca), "_", simplify = TRUE)[,2]
Time <- str_split(rownames(TE_pca), "_", simplify = TRUE)[,3]

TE_pca <- data.frame(cbind(Community, Replicate, Time, TE_pca[, apply(TE_pca[,4:ncol(TE_pca)], 2, sum)>0]))
TE_pca <- TE_pca[apply(TE_pca[,4:ncol(TE_pca)], 1, sum)>0, ]
```

### Select T2 = 7 days of growth

```{r}
TE_pca <- TE_pca[TE_pca$Time == "T2", ]

pca <- PCA(log(TE_pca[, 4:ncol(TE_pca)] + 1), 
           quali.sup = c(1:3),
           scale.unit = FALSE, 
           graph = FALSE
           )

fviz_pca_ind(pca,
             habillage = TE_pca$Community, 
             geom = c("point", "text")
             )
```

# Guild-based Microbiome Classification (Functional Profiling)  

Algorithm:  

- Average replicate
- Make table with rows = bacteria, columns = pathways
- PCA
- HCPC (Hierarchical clustering on the principal components): creates guilds of bacteria that prioritize the same pathways


## TE- based functional profiling

```{r}
# Community 1 average
TE_pca <- TE_syncom[, colnames(TE_syncom)%in%c("X1_1_T2_TE", "X1_2_T2_TE", "X2_1_T2_TE", "X2_2_T2_TE")]

TE_pca <- apply(TE_pca, 1, mean)

TE_pca <- data.frame(cbind(TE_syncom[, c("kegg", "genus_strain")], TE = TE_pca))

TE_pca <- reshape(TE_pca,
        direction = "wide", 
        idvar = "genus_strain", 
        timevar = "kegg")

TE_pca[is.na(TE_pca)] <- 0

rownames(TE_pca) <- TE_pca$genus_strain
TE_pca <- TE_pca[, !colnames(TE_pca) == "genus_strain"]

TE_pca <- TE_pca[apply(TE_pca, 1, sum)!=0, ]
TE_pca <- TE_pca[, apply(TE_pca, 2, sum)!=0]

colnames(TE_pca) <- gsub("TE.", "", colnames(TE_pca))
colnames(TE_pca) <- str_split(colnames(TE_pca), "\\[", simplify = TRUE)[,1]

```

PCA

```{r,  fig.height=4.5, fig.width=6.5}
pca <- PCA(log(TE_pca + 1),
           scale.unit = FALSE,
           graph = FALSE,
           ncp = 5
           )

var_explained <- data.frame(pca$eig)
var_explained$PC <- rownames(var_explained)
var_explained$PC <- ordered(var_explained$PC, levels = unique(var_explained$PC))

ggplot(var_explained, aes(x = PC, y = cumulative.percentage.of.variance)) +
  geom_bar(stat = "identity") +
  geom_vline(xintercept = 5.5, col = "red")
ggplot(var_explained, aes(x = PC, y = percentage.of.variance)) +
  geom_bar(stat = "identity") +
    geom_vline(xintercept = 5.5, col = "red")
```

88% var explained by 5 comp. Clustering stable from 3 components (73% explained)

### Figure 1b

```{r,  fig.height=4.5, fig.width=6.5}
fviz_pca_ind(pca,
             geom = c("point", "text"), 
             repel = TRUE
             )

fviz_pca_biplot(pca,
             geom = c("point", "text"), 
             repel = TRUE, 
             select.var = list(contrib=15), 
             col.ind = "grey", 
             col.var = "black"
             )


TE_clust <- HCPC(pca, 
                 nb.clust = -1, 
                 metric = "euclidean", 
                 method = "ward"
                 )

options(ggrepel.max.overlaps = Inf)

image <- fviz_cluster(TE_clust, repel = TRUE, labelsize = 13, 
                      ggtheme = theme(panel.background = element_rect(fill = "white"), panel.grid.major = element_line(colour = "gray92"), panel.grid.minor = element_line(colour = "gray92")) 
                              ) 
image
# library(svglite)
# setwd("/home/moyne/Moyne_disk/Soil_project/Final_datasets/")
# ggsave(file="guild_clusterplot_white.svg", plot=image, height = 4.5, width = 6.5, dpi = 600)
```

```{r}
tree <- hclust(dist(pca$ind$coord, method = "euclidean"), method = "ward.D2")
cutree(tree, k = 6)
plot(x = tree, labels =  row.names(tree), cex = 1)
plot(x = tree, labels =  row.names(tree), cex = 1, hang = -1)
# 
# setwd("/home/moyne/Moyne_disk/Soil_project/Final_datasets")
# svg("guilds_TE_tree_5PCs.svg", height = 5, width = 6)
# plot(x = tree, labels =  row.names(tree), cex = 1, hang = -1)
# dev.off()
```

# Guilds predict bacterial competition   

```{r, fig.height=7, fig.width=8}
hclustfunc <- function(x) hclust(x, method="ward.D2") # clustering ward
distfunc <- function(x) dist(x, method="euclidean")   # distances euclidiennes 

distmat <- as.data.frame(as.matrix(dist(pca$ind$coord, method = "euclidean")))


#ncp = 5
distmat <- distmat[rev(c("Paenibacillus", "Chitinophaga", "Mucilaginibacter",  "Bosea",  "Methylobacterium",
                        "Rhizobium", "Bradyrhizobium", "Rhodococcus", "Variovorax",
                                                     "Burkholderia", "Niastella", 
                                                    "Lysobacter", "Arthrobacter", "Brevibacillus"
                                                )),
        rev(c("Paenibacillus", "Chitinophaga", "Mucilaginibacter",  "Bosea",  "Methylobacterium",
                        "Rhizobium", "Bradyrhizobium", "Rhodococcus", "Variovorax",
                                                     "Burkholderia", "Niastella", 
                                                    "Lysobacter", "Arthrobacter", "Brevibacillus"
                                                ))]

library(RColorBrewer)

distmat[distmat == 0] <- NA

#  setwd("/home/moyne/Moyne_disk/Soil_project/Final_datasets")
#  svg("TE_distances_5PCs.svg")
heatmap.2(as.matrix(distmat),
          trace = "none",
          Rowv = FALSE,
          Colv = FALSE,
          margins = c(10,10),
          col=redblue,
          scale = "row",
          # labCol=as.expression(lapply(colnames(score), function(a) bquote(italic(.(a))))),
          # labRow=as.expression(lapply(rownames(score), function(a) bquote(italic(.(a))))),
          na.color = "black"
)
# 
# dev.off()
```

## COMPETITION SCORE 

```{r, fig.height=7, fig.width=7}

# -(Z-score of distances)
score <- data.frame(t(apply(distmat, 1, function(x) - (x - mean(x, na.rm = TRUE)) / sd(x, na.rm = TRUE))))

# set up as 0 if abundance was <0.5%
below_0.5 <- c("Methylobacterium", "Marmoricola", "Mycobacterium", "Bosea", "Variovorax", "Rhodococcus")
score[, colnames(score) %in% below_0.5][score[, colnames(score) %in%below_0.5]>0] <- 0

apply(score, 1, function(x)sum(x, na.rm = TRUE))
apply(score, 1, function(x)summary(x, na.rm = TRUE))
score

cool = rainbow(50, start=rgb2hsv(col2rgb('cyan'))[1], end=rgb2hsv(col2rgb('blue'))[1])
warm = rainbow(50, start=rgb2hsv(col2rgb('red'))[1], end=rgb2hsv(col2rgb('yellow'))[1])
cols = c(rev(cool), "lightcyan", rev(warm))
cols = c(rev(cool), "lemonchiffon", rev(warm))
cols = c(rev(cool),  rgb(124, 248, 248, maxColorValue=255, alpha=255),  rev(warm))

mypalette <- colorRampPalette(cols)(101)

 # setwd("/home/moyne/Moyne_disk/Soil_project/Final_datasets")
 # svg("TE_scores_5PCs_lightcyan_lightgray.svg")
heatmap.2(as.matrix(score), 
          trace = "none", 
          Rowv = FALSE,
          Colv = FALSE, 
          margins = c(10,10), 
         col = mypalette, 
          labCol=as.expression(lapply(colnames(score), function(a) bquote(italic(.(a))))),
          labRow=as.expression(lapply(rownames(score), function(a) bquote(italic(.(a))))),
         scale = "none", 
         #side = -1, 
         na.color = "gray90",
         sepwidth=c(0.01,0.01),
          rowsep=c(0:16), 
                   colsep=c(0,14), 
         sepcolor = "white"
)
# dev.off()
```

```{r, fig.height=7, fig.width=7}

# Binary (competition yes/no)

score[score>0] <- 1
# score[score==0] <- 0.5
score[score<0] <- 0

heatmap.2(as.matrix(score), 
          trace = "none", 
          Rowv = FALSE,
          Colv = FALSE, 
          margins = c(10,10), 
          col=bluered, 
         # col = rev(brewer.pal(11,"Spectral")),
         # col = mypalette, 
         scale = "none", 
          na.color = "black"
)
```
